✨ feat: transforming molecule into organism, solving filter cleaning button
